requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
dateparser==1.2.0
Django==5.0.6
fonttools==4.53.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
matplotlib==3.9.0
multidict==6.0.5
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycryptodome==3.20.0
pyparsing==3.1.2
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
websockets==12.0
yarl==1.9.4
